requirements.txt
asgiref==3.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.1.0
cycler==0.11.0
distlib==0.3.6
Django==5.2.6
django-admin-interface==0.28.8
django-appconf==1.1.0
django-autocomplete-light==3.9.7
django-colorfield==0.11.0
django-grappelli==4.0.1
django-select2==8.4.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
filelock==3.10.7
fonttools==4.42.1
idna==3.6
kiwisolver==1.4.5
matplotlib==3.7.2
networkx==3.4.2
numpy==1.25.2
openpyxl==3.1.5
packaging==23.1
Pillow==10.0.0
platformdirs==3.2.0
pyparsing==3.0.9
python-barcode==0.16.1
python-dateutil==2.8.2
python-slugify==8.0.4
reportlab==4.4.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
urllib3==2.1.0
virtualenv==20.21.0
gunicorn